SON HAL

+++ IMDBMovieProject.WebApi/Areas/Admin/Controllers/CategoriesController.cs
@@ -1,19 +1,16 @@
 ﻿using IMDBMovieProject.DataAccess.Data;
 using IMDBMovieProject.Entities.Entities;
-using IMDBMovieProject.WebApi.Utils;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
 namespace IMDBMovieProject.WebApi.Areas.Admin.Controllers
 {
-    [Area("Admin"),Authorize(Policy = "AdminPolicy")]
-    
+    [Area("Admin")]
+    [Authorize(Policy = "AdminPolicy")]
     public class CategoriesController : Controller
     {
         private readonly DataBaseContext _context;
@@ -29,40 +26,27 @@ namespace IMDBMovieProject.WebApi.Areas.Admin.Controllers
             return View(await _context.Categories.ToListAsync());
         }
 
-        // GET: Admin/Categories/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var category = await _context.Categories
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (category == null)
-            {
-                return NotFound();
-            }
-
-            return View(category);
-        }
-
         // GET: Admin/Categories/Create
         public IActionResult Create()
         {
             ViewBag.Kategoriler = new SelectList(_context.Categories, "Id", "Name");
             return View();
         }
+
+        // POST: Admin/Categories/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Category category)
         {
             if (ModelState.IsValid)
-            {                
-                await _context.AddAsync(category);
+            {
+                await _context.Categories.AddAsync(category);
                 await _context.SaveChangesAsync();
+
+                TempData["SuccessMessage"] = "Kategori başarıyla eklendi.";
                 return RedirectToAction(nameof(Index));
             }
+
             ViewBag.Kategoriler = new SelectList(_context.Categories, "Id", "Name");
             return View(category);
         }
@@ -71,48 +55,33 @@ namespace IMDBMovieProject.WebApi.Areas.Admin.Controllers
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
-            {
                 return NotFound();
-            }
 
             var category = await _context.Categories.FindAsync(id);
             if (category == null)
-            {
                 return NotFound();
-            }
+
             ViewBag.Kategoriler = new SelectList(_context.Categories, "Id", "Name");
             return View(category);
         }
 
+        // POST: Admin/Categories/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, Category category)
         {
             if (id != category.Id)
-            {
                 return NotFound();
-            }
 
             if (ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(category);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!CategoryExists(category.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
+                _context.Update(category);
+                await _context.SaveChangesAsync();
+
+                TempData["SuccessMessage"] = "Kategori başarıyla güncellendi.";
                 return RedirectToAction(nameof(Index));
             }
+
             ViewBag.Kategoriler = new SelectList(_context.Categories, "Id", "Name");
             return View(category);
         }
@@ -121,16 +90,17 @@ namespace IMDBMovieProject.WebApi.Areas.Admin.Controllers
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
-            {
                 return NotFound();
-            }
 
             var category = await _context.Categories
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (category == null)
-            {
                 return NotFound();
-            }
+
+            // 🔒 Bu kategoriye bağlı haber var mı?
+            bool hasNews = await _context.News.AnyAsync(n => n.CategoryId == id);
+            ViewBag.HasNews = hasNews;
 
             return View(category);
         }
@@ -140,19 +110,26 @@ namespace IMDBMovieProject.WebApi.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            // 🔒 Güvenlik için tekrar kontrol
+            bool hasNews = await _context.News.AnyAsync(n => n.CategoryId == id);
+
+            if (hasNews)
+            {
+                TempData["ErrorMessage"] =
+                    "Bu kategoriye bağlı haberler bulunduğu için silme işlemi gerçekleştirilemedi.";
+                return RedirectToAction(nameof(Index));
+            }
+
             var category = await _context.Categories.FindAsync(id);
             if (category != null)
             {
                 _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+
+                TempData["SuccessMessage"] = "Kategori başarıyla silindi.";
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool CategoryExists(int id)
-        {
-            return _context.Categories.Any(e => e.Id == id);
         }
     }
 }
